Refactor FieldOptionsList
